+++ src/main/java/com/badbones69/crazycrates/config/impl/ConfigKeys.java
@@ -5,6 +5,8 @@ import ch.jalu.configme.SettingsHolder;
 import ch.jalu.configme.configurationdata.CommentsConfiguration;
 import ch.jalu.configme.properties.Property;
 import com.badbones69.crazycrates.api.enums.State;
+import com.ryderbelserion.vital.paper.api.enums.Support;
+
 import java.util.Collections;
 import java.util.List;
 import static ch.jalu.configme.properties.PropertyInitializer.newBeanProperty;
@@ -114,6 +116,16 @@ public class ConfigKeys implements SettingsHolder {
             ""
     })
     public static final Property<State> message_state = newBeanProperty(State.class, "root.message-state", State.send_message);
+
+    @Comment({
+            "A list of available hologram plugins:",
+            " -> DecentHolograms",
+            " -> FancyHolograms",
+            " -> CMI",
+            "",
+            "If the option is set to blank, it'll pick whatever plugin it feels like picking."
+    })
+    public static final Property<String> hologram_plugin = newProperty("root.hologram-plugin", "");
 
     //@Comment({
     //        "Sends anonymous statistics about how the plugin is used to bstats.org.",