+++ CHANGELOG.md
@@ -1,2 +1,3 @@
-### Fixed:
-- Fixed a class cast exception with the ItemBuilder when using `Data:` or `DisplayData:`
+### Changes:
+- Added a new config option, `hologram-plugin` which lets you pick a hologram plugin to prioritize
+  - If you leave it empty, it'll automatically pick one.